Reimplement AttributedString with a BTree (#18)
A BTree has much better performance characteristics than a Vec for text editing. A Tree has O(log N) complexity for lookups (by character index or by element UID), inserts, and deletes. A Vec has O(N) complexity for inserts and deletes. 

As a result of the change, large AttributedStrings should perform much better than they did with a Vec-based implementation.

Also includes a change to the raw serializer so that it returns an error on failure instead of panicking.