add Counter CRDT type
A `Counter` is a mutable 64-bit float CRDT. It has a single function, `increment`, which increases or decreases the counter's value. Counters automatically discard duplicate operations.

To satisty the requirements needed to reject duplicates, each counter must store the replica of the most recent edit for each editing site. Therefore the space complexity of a counter is O(S), where S is the number of sites which have edited the counter. Time complexity for each increment is O(1).