Register CRDT
The Register CRDT replaces the Atom CRDT. It performs the same function,
as a container for a single value. The name has been updated to reflect
terminology used in the Shapiro et al paper.

The Register CRDT can be serialized with any serde serializer.
It does not contain any custom ser/de code.