improve sequence UID performance (#51)
The old BASE_LEVEL, 3, was too short to be useful in practice (it typically only allowed 1 or 2 elements). The new BASE_LEVEL, 9, should allow ~146 elements. In general, starting at 9 means that UIDs will require fewer levels than before, taking up less space.

I also decreased the BOUNDARY from 10 to 7 because I think a smaller boundary is more efficient for the typical use case, text insertion at the end of the sequence, and not much less efficient for the atypical use case, editing text in the middle of sequence.